Extract per-model agent state into standalone structs
Move each operational model's nested State struct into its own
XModelState.hpp and compose them in the new OperationalModelState
variant header. GenericAgent now depends only on the state headers
instead of every model class header; the ModelAgentState concept,
ModelTypeOf() and the new Pos() accessors move next to the variant.
Each model keeps a `using State = XModelState;` alias so all call
sites and the Python bindings stay source compatible.

No behavior change: state fields, defaults and formatter output are
byte-identical.
