+++ libsimulator/src/GenericAgent.hpp
@@ -1,43 +1,16 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 #pragma once
-#include "AnticipationVelocityModel.hpp"
-#include "CollisionFreeSpeedModel.hpp"
-#include "CollisionFreeSpeedModelV2.hpp"
-#include "CollisionFreeSpeedModelV3.hpp"
-#include "GeneralizedCentrifugalForceModel.hpp"
-#include "OperationalModel.hpp"
-#include "OperationalModels/CustomModel/CustomModel.hpp"
-#include "OperationalModels/OperationalModelType.hpp"
+#include "OperationalModels/OperationalModelState.hpp"
 #include "Point.hpp"
-#include "SocialForceModel.hpp"
 #include "UniqueID.hpp"
-#include "Visitor.hpp"
-#include "WarpDriver/WarpDriverModel.hpp"
 
 #include <fmt/core.h>
 
-#include <concepts>
 #include <deque>
 #include <utility>
 #include <variant>
 class Journey;
 class BaseStage;
-
-/// Agent position is owned by the per-model agent state. Every alternative of
-/// GenericAgent::ModelState must satisfy this concept; the framework accesses the
-/// position type-erased through GenericAgent::position().
-template <typename T>
-concept ModelAgentState = requires(T t) {
-    // position() hands out mutable Point& into the state, so a convertible or const member
-    // is not enough.
-    { t.position } -> std::same_as<Point&>;
-};
-
-template <typename Variant>
-inline constexpr bool EachAlternativeIsModelAgentState = false;
-template <typename... Ts>
-inline constexpr bool EachAlternativeIsModelAgentState<std::variant<Ts...>> =
-    (ModelAgentState<Ts> && ...);
 
 struct GenericAgent {
     using ID = jps::UniqueID<GenericAgent>;
@@ -50,28 +23,11 @@ struct GenericAgent {
     Point destination{};
     Point target{};
 
-    using ModelState = std::variant<
-        GeneralizedCentrifugalForceModel::State,
-        CollisionFreeSpeedModel::State,
-        CollisionFreeSpeedModelV2::State,
-        CollisionFreeSpeedModelV3::State,
-        AnticipationVelocityModel::State,
-        SocialForceModel::State,
-        WarpDriverModel::State,
-        CustomModel::State>;
-    static_assert(
-        EachAlternativeIsModelAgentState<ModelState>,
-        "Every agent model state must provide a 'Point position' member");
+    using ModelState = OperationalModelState;
     ModelState model{};
 
-    Point& position()
-    {
-        return std::visit([](auto& m) -> Point& { return m.position; }, model);
-    }
-    const Point& position() const
-    {
-        return std::visit([](const auto& m) -> const Point& { return m.position; }, model);
-    }
+    Point& position() { return Pos(model); }
+    const Point& position() const { return Pos(model); }
 
     GenericAgent(
         ID id_,
@@ -87,34 +43,6 @@ struct GenericAgent {
         target = position();
     }
 };
-
-/// Maps agent model data to the operational model type it belongs to. Kept
-/// exhaustive on purpose: adding a model type will not compile until the
-/// mapping is extended.
-inline OperationalModelType ModelTypeOf(const GenericAgent::ModelState& model)
-{
-    return std::visit(
-        overloaded{
-            [](const GeneralizedCentrifugalForceModel::State&) {
-                return OperationalModelType::GENERALIZED_CENTRIFUGAL_FORCE;
-            },
-            [](const CollisionFreeSpeedModel::State&) {
-                return OperationalModelType::COLLISION_FREE_SPEED;
-            },
-            [](const CollisionFreeSpeedModelV2::State&) {
-                return OperationalModelType::COLLISION_FREE_SPEED_V2;
-            },
-            [](const CollisionFreeSpeedModelV3::State&) {
-                return OperationalModelType::COLLISION_FREE_SPEED_V3;
-            },
-            [](const AnticipationVelocityModel::State&) {
-                return OperationalModelType::ANTICIPATION_VELOCITY_MODEL;
-            },
-            [](const SocialForceModel::State&) { return OperationalModelType::SOCIAL_FORCE; },
-            [](const WarpDriverModel::State&) { return OperationalModelType::WARP_DRIVER; },
-            [](const CustomModel::State&) { return OperationalModelType::CUSTOM_MODEL; }},
-        model);
-}
 
 template <class Agent>
 using AgentContainer = std::deque<Agent>;
