+++ libsimulator/src/OperationalModels/CollisionFreeSpeedModelV2/CMakeLists.txt
@@ -2,5 +2,6 @@
 target_sources(simulator PRIVATE
     CollisionFreeSpeedModelV2.cpp
     CollisionFreeSpeedModelV2.hpp
+    CollisionFreeSpeedModelV2State.hpp
 )
 target_include_directories(simulator PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
