+++ libsimulator/src/OperationalModels/CollisionFreeSpeedModelV3/CMakeLists.txt
@@ -2,5 +2,6 @@
 target_sources(simulator PRIVATE
     CollisionFreeSpeedModelV3.cpp
     CollisionFreeSpeedModelV3.hpp
+    CollisionFreeSpeedModelV3State.hpp
 )
 target_include_directories(simulator PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
