+++ libsimulator/src/OperationalModels/CustomModel/CMakeLists.txt
@@ -1,4 +1,5 @@
 target_sources(simulator PRIVATE
     CustomModel.hpp
+    CustomModelState.hpp
     FormatAny.hpp
 )
