+++ libsimulator/src/OperationalModels/GeneralizedCentrifugalForceModel/CMakeLists.txt
@@ -2,5 +2,6 @@
 target_sources(simulator PRIVATE
     GeneralizedCentrifugalForceModel.cpp
     GeneralizedCentrifugalForceModel.hpp
+    GeneralizedCentrifugalForceModelState.hpp
 )
 target_include_directories(simulator PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
