+++ libsimulator/src/OperationalModels/SocialForceModel/CMakeLists.txt
@@ -2,5 +2,6 @@
 target_sources(simulator PRIVATE
     SocialForceModel.cpp
     SocialForceModel.hpp
+    SocialForceModelState.hpp
 )
 target_include_directories(simulator PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
